Done with Login/Signup Authentication

+++ frontend/src/components/WorkoutDetails.js
@@ -1,16 +1,25 @@
 
 import {useWorkoutsContext} from '../hooks/useWorkoutsContext'
 import {formatDistanceToNow} from 'date-fns/formatDistanceToNow'
+import { useAuthContext } from '../hooks/useAuthContext'
 
 
 
 const WorkoutDetails = ({workout}) => {
     const {dispatch} = useWorkoutsContext()
+    const {user} = useAuthContext()
     
     // Delete function
     const handleClickDel = async () => {
+
+        if (!user){
+            return
+        }
         const response = await fetch('/api/workouts/' + workout._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
